MYR-585 Split copy.go by audience to clear the 300-line ceiling
Review flagged that copy.go had reached 429 lines against CLAUDE.md's 300-line
cap, and that this PR deepened it by 59. Splitting by AUDIENCE rather than
arbitrarily: copy_rider.go answers what the passenger needs to know,
copy_owner.go what the person whose car it is needs to know, and copy.go keeps
the payload policy, the status/stamp vocabulary and the two interpolation
helpers that enforce the policy mechanically. 150 / 113 / 197 lines.

This also fixes the doc-comment attachment drift the review found, by
construction: statusAlert's and ownerStatusAlert's doc blocks had been absorbed
into the declarations that happened to follow them (ownerCancelled and
riderCancelKind), and each is now attached to the function it describes.

Two duplicated body literals ('Try booking another car.') became the shared
bodyTryAnotherCar, joining the constants the rest of the file already uses.
